+++ v2023/src/main.rs
@@ -25,6 +25,7 @@ fn main() {
         "3.2" => stars::day_03::adjacent_gear_ratios(input),
         "4.1" => stars::day_04::scratchcard_points(input),
         "4.2" => stars::day_04::scratchcard_replay(input),
+        "5.1" => stars::day_05::seed_to_location_min(input),
         _ => {
             eprintln!("error: star '{}' does not yet shine", args.star);
             std::process::exit(1);