+++ CHANGELOG.md
@@ -8,6 +8,7 @@ the format is inspired by [keep a changelog][changelog] with entries logged as
 ## [unreleased]
 
 - feat(2023): include a helper function to parse numbers in a string
+- feat(2023.5.1): find the seed that maps to the lowest location
 
 ## [0.2023.14]